inference-server/requirements.txt
Flask==3.0.3
Requests==2.31.0
torch==2.2.2
torchvision==0.17.2
Werkzeug==3.0.2
scikit-learn==1.3.2
scipy==1.13.0
pillow==10.3.0
joblib==1.4.0
matplotlib-inline==0.1.6
prometheus-client==0.20.0
prometheus-flask-exporter==0.23.0
flasgger==0.9.7.1
python-dotenv==1.0.1
requests-toolbelt==1.0.0

#flask-cors
#flask_httpauth
#gunicorn